Multiple fixes - File restructuring - Fetch Optimization
