+++ components/Creator/CreateUI.tsx
@@ -1,135 +1,23 @@
-import React, { useState, useMemo } from 'react';
+import React from 'react';
 import { motion } from 'framer-motion';
-
-export interface videoData {
-    author_name: string;
-    author_url: string;
-    height: number;
-    html: string;
-    provider_name: string;
-    provider_url: string;
-    thumbnail_height: number;
-    thumbnail_url: string;
-    thumbnail_width: number;
-    title: string;
-    type: string;
-    version: string;
-    width: number;
-}
+import { useVideoInfo } from './hooks/useVideoInfo';
+import { useTags } from './hooks/useTags';
+import useFormSubmit from './hooks/useFormSubmit';
 
 function CreateUI() {
-    const [url, setUrl] = useState('');
-    const [tags, setTags] = useState('');
-    const [isPresent, setIsPresent] = useState(false);
-    const [isSubmittable, setIsSubmittable] = useState(false);
+    const { url, focussedVideo, isPresent, fetchAndSetVideoInfo } = useVideoInfo();
+    const { tags, setTagsFromInput } = useTags();
+    const { isSubmittable, handleSubmit } = useFormSubmit(url, tags, focussedVideo, isPresent);
 
-    const [focussedVideo, setFocussedVideo] = useState({} as videoData);
-
-    const validateUrl = (url: string) => {
-        const re = /^(https?\:\/\/)?(www\.)?(youtube\.com|youtu\.?be)\/.+$/;
-        return re.test(url);
-    };
-
-    const fetchVideoInfo = async (url: string) => {
-        const oembedUrl = `https://www.youtube.com/oembed?url=${url}&format=json`;
-        const data = await fetch(oembedUrl).then(res => res.json());
-        data.html = data.html.replace(/width="\d+"/, 'width="100%"').replace(/height="\d+"/, 'height="113"');
-        return data;
-    };
-
-    const unBlurUrlFieldHandler = async (e: React.FocusEvent<HTMLInputElement>) => {
+    const onBlurUrlFieldHandler = async (e: React.FocusEvent<HTMLInputElement>) => {
         const url = e.target.value;
-
-        if (!url) {
-            setFocussedVideo({} as videoData);
-            setIsPresent(false);
-            return;
-        }
-
-        if (validateUrl(url)) {
-            setUrl(url);
-            checkPresence(url);
-            const data = await fetchVideoInfo(url);
-            setFocussedVideo(data);
-        }
+        await fetchAndSetVideoInfo(url);
     };
 
-    const unBlurTagsFieldHandler = (e: React.FocusEvent<HTMLInputElement>) => {
+    const onBlurTagsFieldHandler = (e: React.FocusEvent<HTMLInputElement>) => {
         const tags = e.target.value;
-
-        if (!tags) {
-            setTags('');
-            return;
-        }
-
-        setTags(tags);
+        setTagsFromInput(tags);
     };
-
-    const checkPresence = async (url: string) => {
-        const res = await fetch('/api/presenceCheck', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({ url }),
-        });
-        const presence = await res.json();
-
-        setIsPresent(presence.status);
-    };
-
-    const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-
-        if (!url) {
-            return;
-        }
-
-        if (!validateUrl(url)) {
-            return;
-        }
-
-        // console.log(url, tags);
-
-        // make a call to the create endpoint
-        const res = await fetch('/api/create', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            // add all data from focusVideo and the tags and url to the body
-            body: JSON.stringify({
-                ...focussedVideo,
-                tags,
-                url,
-            }),
-        });
-
-        const data = await res.json();
-
-        // console.log(data);
-        checkPresence(url);
-    };
-
-    const handleDebugSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-
-        // Change this to the endpoint you want to debug
-        const debug_call = await fetch('/api/calculateColor');
-        // console.log(debug_call)
-    };
-
-    useMemo(() => {
-        // console.log('useMemo called', url, tags, isPresent);
-
-        if (url && tags && !isPresent) {
-            // console.log('setting isSubmittable to true');
-            setIsSubmittable(true);
-        } else {
-            // console.log('setting isSubmittable to false');
-            setIsSubmittable(false);
-        }
-    }, [url, tags, isPresent]);
 
     return (
         <main className="flex h-full w-full flex-col gap-2 p-6">
@@ -146,7 +34,7 @@ function CreateUI() {
                                 name="url"
                                 id="url"
                                 className="rounded bg-indigo-800/50 p-1"
-                                onBlur={unBlurUrlFieldHandler}
+                                onBlur={onBlurUrlFieldHandler}
                             />
                             <label htmlFor="tags" className="text-center align-middle">
                                 Tags
@@ -156,46 +44,34 @@ function CreateUI() {
                                 name="tags"
                                 id="tags"
                                 className="rounded bg-indigo-800/50 p-1"
-                                onBlur={unBlurTagsFieldHandler}
+                                onBlur={onBlurTagsFieldHandler}
                             />
                         </fieldset>
                         <motion.button
-                            className={`place-self-center rounded p-1 text-white
-    							${isSubmittable ? 'bg-indigo-800/50' : 'cursor-not-allowed bg-gray-700/25'}`}
+                            className={`place-self-center rounded p-1 text-white ${isSubmittable ? 'bg-indigo-800/50' : 'cursor-not-allowed bg-gray-700/25'}`}
                             disabled={!isSubmittable}
                             whileHover={{ scale: 1.05 }}
                             whileTap={{ scale: 0.95 }}
                         >
                             Submit
                         </motion.button>
-                        {/* <motion.button
-							className={`place-self-center rounded p-1 text-white bg-red-500`}
-							whileHover={{ scale: 1.05 }}
-							whileTap={{ scale: 0.95 }}
-							//@ts-ignore
-							onClick={handleDebugSubmit}
-						>
-							Debug
-						</motion.button> */}
                     </form>
                 </div>
             </section>
-            <section className="flex h-1/3 w-full flex-col place-items-center justify-center rounded bg-indigo-500/5 p-2">
+            <section
+                className="flex h-1/3 w-full flex-col place-items-center justify-center rounded bg-indigo-500/5 p-2">
                 <span className="order-first text-center text-xl font-bold">You're looking at:</span>
-                <span className="w-full truncate text-center text-lg">{focussedVideo.title}</span>
+                <span className="w-full truncate text-center text-lg">{focussedVideo?.title}</span>
                 <div className="w-fit p-2">
                     {focussedVideo && (
-                        /* insert iframe from focussedVideo.html here */
                         <div
                             className="h-full w-full"
-                            dangerouslySetInnerHTML={{
-                                __html: focussedVideo.html,
-                            }}
+                            dangerouslySetInnerHTML={{ __html: focussedVideo.html }}
                         />
                     )}
                 </div>
                 <span className="text-md w-full flex-grow truncate text-center text-red-600">
-                    {isPresent && "... It's already in here!"}
+                    {isPresent && '... It\'s already in here!'}
                 </span>
             </section>
         </main>
