+++ components/Creator/hooks/useTags.ts
@@ -1,0 +1,14 @@
+import { useState } from 'react';
+
+export const useTags = () => {
+    const [tags, setTags] = useState('');
+
+    const setTagsFromInput = (input: string) => {
+        setTags(input || '');
+    };
+
+    return {
+        tags,
+        setTagsFromInput,
+    };
+};
