+++ components/Creator/hooks/utils.ts
@@ -1,0 +1,4 @@
+export const validateUrl = (url: string) => {
+    const re = /^(https?:\/\/)?(www\.)?(youtube\.com|youtu\.?be)\/.+$/;
+    return re.test(url);
+};
