+++ components/Creator/types/IVideoData.ts
@@ -1,0 +1,15 @@
+export interface IVideoData {
+    author_name: string;
+    author_url: string;
+    height: number;
+    html: string;
+    provider_name: string;
+    provider_url: string;
+    thumbnail_height: number;
+    thumbnail_url: string;
+    thumbnail_width: number;
+    title: string;
+    type: string;
+    version: string;
+    width: number;
+}
