+++ components/Player/types/states.ts
