+++ components/Viewer/NotificationButton.tsx
@@ -62,13 +62,13 @@ const NotificationButton: React.FC<NotificationButtonProps> = ({
 							xmlns='http://www.w3.org/2000/svg'
 							fill='none'
 							viewBox='0 0 24 24'
-							stroke-width='1.5'
+							strokeWidth='1.5'
 							stroke='currentColor'
 							className='h-5 w-full'
 						>
 							<path
-								stroke-linecap='round'
-								stroke-linejoin='round'
+								strokeLinecap='round'
+								strokeLinejoin='round'
 								d='M4.5 12.75l6 6 9-13.5'
 							/>
 						</svg>
