+++ components/Viewer/hooks/useItems.ts
@@ -1,7 +1,7 @@
 // useItems.ts
 import useSWR from 'swr';
 import { track, tag } from '@prisma/client';
-import { useFilterState } from '../../contexts/FilterStateProvider';
+import { useFilter } from '../../contexts/FilterStateProvider';
 
 const fetcher = (url: string, body?: any) =>
     fetch(url, {
@@ -13,25 +13,25 @@ const fetcher = (url: string, body?: any) =>
     }).then((res) => res.json());
 
 export function useItems(type: string) {
-    const { filterState } = useFilterState();
+    const { filter } = useFilter();
 
     let route = '/api/list';
     if (type === 'new') {
-        route = '/api/new';
+        route = '/api/viewer/new';
     } else if (type === 'trend') {
-        route = '/api/list';
+        route = '/api/viewer/list';
     } else if (type === 'filter') {
-        route = '/api/filter';
+        route = '/api/viewer/filter';
     } else if (type === 'owned') {
-        route = '/api/list';
+        route = '/api/viewer/list';
     }
 
     const regFetcher = (route: string) => (() => fetcher(route));
     const FilteredFetcher = (route: string, filter: string[]) => (() => fetcher(route, { filter: filter }));
 
     const { data, error, isLoading, mutate } = useSWR<track & { tags: tag[] }[], Error>(
-        type === 'filter' ? ['/api/filter', filterState] : route,
-        type === 'filter' ? FilteredFetcher(route, filterState) : regFetcher(route),
+        type === 'filter' ? ['/api/viewer/filter', filter] : route,
+        type === 'filter' ? FilteredFetcher(route, filter) : regFetcher(route),
         {
             revalidateOnFocus: false,
             refreshInterval: 1000 * 60 * 20,
