--- components/mixerComponents [DEPRECATED]/YTPlayer.module.css
@@ -1,5 +1,0 @@
-.slider {
-    appearance: slider-vertical;
-    width: 5px;
-    height: 125px;
-}
