--- components/utils/VolumeSlider.module.css
@@ -1,75 +1,0 @@
-/* General styles for the slider container */
-.container {
-    display: flex;
-    max-width: max-content;
-    flex-direction: row;
-    align-items: center;
-    gap: 4px;
-    height: 100%;
-}
-
-/* Styles for the slider input element */
-.slider {
-    -webkit-appearance: slider-vertical;
-    appearance: slider-vertical;
-    width: 5px;
-    /* margin-top: 50px; */
-}
-
-/* ------------------ Firefox ------------------ */
-
-.slider::-moz-range-thumb {
-    appearance: none;
-    width: 16px;
-    height: 16px;
-    background-color: red;
-    border-radius: 50%;
-    border: none;
-    transition: 0.2s ease-in-out;
-}
-
-/* ------------------ Chrome ------------------ */
-
-/* 
-.slider::-webkit-slider-runnable-track {
-    -webkit-appearance: none;
-    appearance: none;
-    width: 100%;
-    height: 5px;
-    background-color: rgb(37, 190, 39);
-    border-radius: 2.5px;
-}
-
-.slider::-webkit-slider-thumb {
-    -webkit-appearance: none;
-    appearance: none;
-    border: 1px solid #000000;
-    height: 36px;
-    width: 16px;
-    border-radius: 3px;
-    background: red;
-    cursor: pointer;
-    margin-top: 50px;
-} */
-
-/* ------------------ MS Edge ------------------ */
-
-.slider::-ms-thumb {
-    width: 16px;
-    height: 16px;
-    background-color: red;
-    border-radius: 50%;
-}
-
-.slider::-ms-track {
-    width: 100%;
-    height: 5px;
-    background: #ccc;
-    border-color: transparent;
-    color: transparent;
-}
-
-/*
-.slider:focus {
-    outline: none;
-} */
