+++ pages/api/creator/presenceCheck.ts
@@ -1,6 +1,6 @@
 /* query the database to check if the url is already present */
 import { NextApiRequest, NextApiResponse } from 'next'
-import { prisma } from './prismaClientProvider';
+import { prisma } from '../../../utils/prismaClientProvider';
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse) {
