+++ pages/api/viewer/list.ts
@@ -1,17 +1,19 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { prisma } from "./prismaClientProvider";
+import { prisma } from "../../../utils/prismaClientProvider";
 
 export default async function handle(req: NextApiRequest, res: NextApiResponse) {
     // Query the database
     const result = await prisma.track.findMany({
-        orderBy: {
-            createdAt: "desc"
-        },
         include: {
             tags: true
         }
     });
 
+    // randomize the result order
+    function randomize() {
+        return Math.random() - 0.5;
+    }
+    result.sort(randomize);
 
     res.status(200).json(result);
 }
